Unify Qwen3 ASR KV cache dtype

+++ src/models/optimized_qwen3_asr/tensors/decode_layer.py
@@ -1134,7 +1134,7 @@ def create_decode_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key="index_copy",
-                spec=TensorSpec(dtype="float32", shape=(1, 8, max_sequence_length, 128)),
+                spec=TensorSpec(dtype="float16", shape=(1, 8, max_sequence_length, 128)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1147,7 +1147,7 @@ def create_decode_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key="index_copy_1",
-                spec=TensorSpec(dtype="float32", shape=(1, 8, max_sequence_length, 128)),
+                spec=TensorSpec(dtype="float16", shape=(1, 8, max_sequence_length, 128)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
