+++ src/models/optimized_qwen3_asr/tensors/model.py
@@ -133,7 +133,7 @@ def create_model_tensors(
     )
     key_caches = tuple(
         _request_state_tensor(
-            "float32",
+            "float16",
             (1, num_key_value_heads, max_sequence_length, head_dim),
             semantic=TensorSemantic.KV_CACHE,
         )
@@ -141,7 +141,7 @@ def create_model_tensors(
     )
     value_caches = tuple(
         _request_state_tensor(
-            "float32",
+            "float16",
             (1, num_key_value_heads, max_sequence_length, head_dim),
             semantic=TensorSemantic.KV_CACHE,
         )
