+++ src/models/optimized_qwen3_asr/tensors/text_layer.py
@@ -1049,7 +1049,7 @@ def create_text_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='index_copy',
-                spec=TensorSpec(dtype='float32', shape=(1, 8, max_sequence_length, 128)),
+                spec=TensorSpec(dtype='float16', shape=(1, 8, max_sequence_length, 128)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
@@ -1062,7 +1062,7 @@ def create_text_layer(
             _declare_tensor(
                 checkpoint_key=None,
                 reference_key='index_copy_1',
-                spec=TensorSpec(dtype='float32', shape=(1, 8, max_sequence_length, 128)),
+                spec=TensorSpec(dtype='float16', shape=(1, 8, max_sequence_length, 128)),
                 layout=CONTIGUOUS_LAYOUT,
                 role=TensorRole.ACTIVATION,
                 memory=MemoryClass.FRAME_WORKSPACE,
